+++ sipXportLib/src/os/OsSysLogTask.cpp
@@ -277,10 +277,10 @@ UtlBoolean OsSysLogTask::handleMessage(OsMsg& eventMessage)
                processConsoleEnable(FALSE);
                break ;
             case OsSysLogMsg::HEAD:
-               processHeadCommand((int) pSysLogMsg->getData());
+               processHeadCommand((intptr_t) pSysLogMsg->getData());
                break ;
             case OsSysLogMsg::TAIL:
-               processTailCommand((int) pSysLogMsg->getData());
+               processTailCommand((intptr_t) pSysLogMsg->getData());
                break ;
             case OsSysLogMsg::ADD_SOCKET:
                data = (char*) pSysLogMsg->getData();
@@ -302,7 +302,7 @@ UtlBoolean OsSysLogTask::handleMessage(OsMsg& eventMessage)
                processSetCallback((OsSysLogCallback) pSysLogMsg->getData());
                break ;
             case OsSysLogMsg::SET_FLUSH_PERIOD:
-               processSetFlushPeriod((int) pSysLogMsg->getData()) ;
+               processSetFlushPeriod((intptr_t) pSysLogMsg->getData()) ;
                break ;
             case OsSysLogMsg::FLUSH_LOG:
                processFlushLog((OsEvent*) pSysLogMsg->getData());
